wrap reading data in reading object

+++ app/src/main/java/com/pnpc/pavement/Reading.java
@@ -8,16 +8,16 @@ import com.google.gson.annotations.SerializedName;
 public class Reading {
 
     @SerializedName("start_lat")
-    public float start_lat;
+    public double start_lat;
 
     @SerializedName("start_lon")
-    public float start_lon;
+    public double start_lon;
 
     @SerializedName("end_lat")
-    public float end_lat;
+    public double end_lat;
 
     @SerializedName("end_lon")
-    public float end_lon;
+    public double end_lon;
 
     @SerializedName("acceleration_x")
     public String acceleration_x;
@@ -37,7 +37,7 @@ public class Reading {
     @SerializedName("end_time")
     public float end_time;
 
-    public Reading(float start_lat, float start_lon, float end_lat, float end_lon, String acceleration_x, String acceleration_y,
+    public Reading(double start_lat, double start_lon, double end_lat, double end_lon, String acceleration_x, String acceleration_y,
                                  String acceleration_z, int ride_id) {
         this.start_lat = start_lat;
         this.start_lon = start_lon;
